Fix precision problems

+++ PR_HW1/GaussianPdf.h
@@ -18,6 +18,6 @@ public:
 	GaussianPdf& operator=(const GaussianPdf& x);
 	~GaussianPdf();
 
-	float GetProbability(const FeatureData& x) const;
+	double GetProbability(const FeatureData& x) const;
 };
 #endif
